refactor: change pkg structure
