--- pokeClient/test/client.go
@@ -1,1 +1,0 @@
-package client
